revamped food input field
