+++ client/src/views/DailyDiary.vue
@@ -87,6 +87,9 @@ function generateDateAndTime() {
   dateAndTime.value = currentdateAndTime;
 }
 
+/**
+ * Inserts the diary entry into the database
+ */
 async function insertNotesToDB() {
   try {
     generateDateAndTime();
@@ -105,6 +108,9 @@ async function insertNotesToDB() {
   }
 }
 
+/**
+ * Retrieves the user's diary from the database
+ */
 async function getUserDiary() {
   try {
     const result = await getDiary(username.value);
